requirements.txt
streamlit
feedparser
PyGithub
google-generativeai
python-dotenv
